requirements.txt
Django==1.11.1
djangorestframework==3.6.3
django-rest-swagger==2.0.5
psycopg2==2.7.1
ipdb
